Aula 2013-03-20 - Uri hierarchy using inheritance support in Java
This is the code as it was at the end of the class. It does support fully the Http and Ftp schemas and no parse for mailto and geo.

The problem of Uri creation is not yet solved.

This code does not compile, because tests were not changed to support the refactoring done!