Aula 2013-05-08 - Inversion of control on key reception. This version is a transitory one at the end of this class, and does not work as expected yet!
CrypraiderSwingViewer receives the key events and notifies the game with the key that keeps them. Caeter asks game for a key on each move.
The game main processing loop (CryptRaider.start method) is:
- showBoard();
- moveActors();
- waitForNextFrame(1000/30);

The game does not work as expected yet because the CrypraiderSwingViewer.show methods is recreating the frame each time it is called.
This version does not work as expected yet. Problems will be solved on next lecture.